Separate test framework into own library (gruetest)
https://github.com/swordglowsblue/gruetest
