+++ src/test/func.ts
@@ -1,8 +1,8 @@
-import { describe, testConsistency } from "../testcore";
-import { wildMagicEffects } from "../../opdefs/dnd5e";
-import { parseExpr, parseExprList } from "../../core/parse";
-import { DiceResult } from "../../core/expressions";
-import { Op } from '../../core/operators'
+import describe, { testConsistency } from "gruetest";
+import { wildMagicEffects } from "../opdefs/dnd5e";
+import { parseExpr, parseExprList } from "../core/parse";
+import { DiceResult } from "../core/expressions";
+import { Op } from '../core/operators'
 
 describe('Dice functions', (it,todo) => {
   it('should properly calculate exploding dice', assert => {
